feat: merge Mein Bereich into unified Dashboard for all roles
MITGLIED_AKTIV/PASSIV now see the full personal dashboard (calendar,
profile, widgets, history) at /dashboard instead of a generic stats page.
Vorstand dashboard remains unchanged. /mein-bereich redirects to /dashboard,
sub-pages (stundenkonto, verfuegbarkeit, einstellungen) stay in place.
